requirements.txt
camel-ai[all]
python-dotenv
fnmatch
abc
typing